+++ python/models/recursos_humanos.py
@@ -43,6 +43,6 @@ class SueldosDeIntegrantes(db.Model,BaseMixin,AuditMixin):
 
     @validates('sueldo')
     def validate_non_negative(self, key, value):
-        if Decimal(value) < 0:
+        if float(value) < 0:
             raise ValueError(f"{key.replace('_',' ').capitalize()} no puede ser negativo")
         return value
